% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake-section-title.R, R/included-datasets.R
\docType{data}
\name{commune_level_data}
\alias{commune_level_data}
\title{1171 entries commune level housing data for Luxemborg
\itemize{
\item year. year of listing
\item locality. local authority (smallest administrative divisions of territory)
\item n_offers. total number of selling offers
\item average_price_nominal_euros. average selling price in nominal euros
\item average_price_m2_nominal_euros. average selling price in square meters in nominal euros
}}
\format{
A data frame with 1171 rows and 5 columns

A data frame with 11 rows and 5 columns

A data frame with 1171 rows and 5 variables:
\describe{
\item{ year }{ 2010 to 2020 }
\item{ locality }{ Communes of Luxembourg }
\item{ n_offers }{ Total offers in that commune for that year }
\item{ average_price_nominal_euros }{ Average selling price of houses in that commune for that year }
\item{ average_price_m2_nominal_euros }{ Average selling price by m2 of houses in that commune for that year }
}
}
\source{
Observatoire de l'Habitat
}
\usage{
data(commune_level_data)

data(commune_level_data)

commune_level_data
}
\description{
data in 'luHousePrices'.
}
\keyword{datasets}
