% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_laspeyeres.R
\name{make_plot}
\alias{make_plot}
\title{Plot price index}
\usage{
make_plot(country_level_data, commune_level_data, commune)
}
\arguments{
\item{country_level_data}{The country level data frame}

\item{commune_level_data}{The commune level data frame}

\item{commune}{The commune to plot}
}
\value{
A ggplot
}
\description{
Plot price index
}
\examples{
\dontrun{
commune_level_data_laspeyeres <- get_laspeyeres(commune_level_data)
make_plot(commune_level_data_laspeyeres, "Luxembourg")
}

}
