% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_laspeyeres.R
\name{get_laspeyeres}
\alias{get_laspeyeres}
\title{Calculate price indices}
\usage{
get_laspeyeres(dataset)
}
\arguments{
\item{dataset}{The dataset for which the index needs to be computed}
}
\value{
The input dataset with added columns: p0, p0_m2, pl and pl_m2. p0 are prices at year 0 (2010), and pl are the index.
}
\description{
Calculate price indices
}
\examples{
\dontrun{ 
commune_level_data <- get_laspeyeres(commune_level_data)
country_level_data <- get_laspeyeres(country_level_data)
}
}
